Make request completion callback configurable
Prepare for future alternate rpc batching and blocking mechanisms by
removing req->wq, and adding req->client_cb().

The unused req->aux is used for an arbitrary argument to req->client_cb.
In p9_client_rpc() it is assigned to point to a wait queue that is
signaled upon completion by req->client_cb() as before.